pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["skwwnl <skwwnl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
tortoise-orm = "0.21.6"
fastapi = {extras = ["standard"], version = "0.115.2"}
uvicorn = "0.31.1"
python-jose = "3.3.0"
aerich = "0.7.2"
aiomysql = "0.2.0"
python-dotenv = "1.0.1"
pydantic = "2.9.2"
pydantic-settings = "2.5.2"

[tool.aerich]
tortoise_orm = "app.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
black = "24.10.0"
flake8 = "7.1.1"
mypy = "1.11.2"
isort = "5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
